project restructure bugfix
