+++ backend-springboot/src/main/java/com/cscd488seniorproject/syllabussyncproject/HealthController.java
@@ -1,4 +1,4 @@
-package com.cscd488seniorproject.syllabussync;
+package com.cscd488seniorproject.syllabussyncproject;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
