+++ backend-springboot/src/main/java/com/cscd488seniorproject/syllabussyncproject/SyllabusSyncApplication.java
@@ -1,4 +1,4 @@
-package com.cscd488seniorproject.syllabussync;
+package com.cscd488seniorproject.syllabussyncproject;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
