+++ backend-springboot/src/main/java/com/cscd488seniorproject/syllabussyncproject/TestEntity.java
@@ -1,4 +1,4 @@
-package com.cscd488seniorproject.syllabussync;
+package com.cscd488seniorproject.syllabussyncproject;
 
 import jakarta.persistence.*;
